Fix mypy error on newer numpy: cast data stats to plain float
np.mean/np.std/stats.skew/stats.kurtosis return numpy scalar types
under recent numpy versions, which mypy correctly flagged as
incompatible with JumpDistribution.initial_guess's float-typed
parameters -- only surfaced in CI on Python 3.9+ where a newer numpy
gets resolved, not in the 3.8 job or in local runs against an older
numpy pin.

+++ src/jump_diffusion/estimation/maximum_likelihood.py
@@ -64,10 +64,10 @@ class JumpDiffusionEstimator(BaseEstimator):
 
         # Calculate basic statistics
         self.n_obs = len(self.increments)
-        self.mean_increment = np.mean(self.increments)
-        self.std_increment = np.std(self.increments)
-        self.skewness = stats.skew(self.increments)
-        self.kurtosis = stats.kurtosis(self.increments)
+        self.mean_increment = float(np.mean(self.increments))
+        self.std_increment = float(np.std(self.increments))
+        self.skewness = float(stats.skew(self.increments))
+        self.kurtosis = float(stats.kurtosis(self.increments))
 
     def log_likelihood(self, params: np.ndarray) -> float:
         """
